requirements.txt
pandas
numpy
tensorflow
matplotlib
pillow
scikit-learn
livelossplot